feat: move crypto to own module
