+++ apps/api/src/middleware/auth.ts
@@ -1,92 +1,14 @@
-import * as crypto from 'crypto';
-
 import { makeLogger } from '../logger';
 import prisma from '../prismaClient';
 import { verifyToken } from '../auth/jwt';
 import { NextFunction, Request, Response } from 'express';
 import { User } from '@prisma/client';
 import { UserAwareRequest } from '../util/typescript';
+import { AuthenticationError } from 'spotify-analytics-errors';
 
 const log = makeLogger(module);
 
-const ITERATIONS = 100000;
-const KEYLEN = 64;
-
-export function generateSalt(): Promise<string> {
-    return new Promise<string>((resolve, reject) => {
-        crypto.randomBytes(KEYLEN, (err, salt) => {
-            if (err) {
-                log.error({ err }, `(${generateSalt.name}) - error generating salt`);
-                return reject(err);
-            }
-
-            return resolve(salt.toString('hex'));
-        });
-    });
-}
-
-export function hashPassword(password: string, salt: string): Promise<string> {
-    return new Promise<string>((resolve, reject) => {
-        crypto.pbkdf2(password, salt, ITERATIONS, KEYLEN, 'sha512', (err, derivedKey) => {
-            if (err) {
-                log.error({ err }, `(${hashPassword.name}) - error hashing`);
-                return reject(err);
-            }
-
-            return resolve(derivedKey.toString('hex'));
-        });
-    });
-}
-
-export function verifyUsernamePassword(username, password, done) {
-    log.info({ username, password }, `(${verifyUsernamePassword.name})`);
-
-    prisma.user
-        .findFirst({ where: { username } })
-        .then(user => {
-            log.info({ username, user }, `(${verifyUsernamePassword.name})`);
-
-            if (!user) {
-                return done(null, false, { message: 'Incorrect credentials provided' });
-            }
-
-            crypto.pbkdf2(password, user?.passwordSalt ?? 'salt', ITERATIONS, KEYLEN, 'sha512', (err, derivedKey) => {
-                if (err) {
-                    log.error({ err }, `(${verifyUsernamePassword.name}) - error hashing`);
-                    return done(err, false, { message: 'Incorrect credentials provided' });
-                }
-
-                if (derivedKey.toString('hex') === user?.passwordHash) {
-                    return done(null, user);
-                }
-
-                log.error({ message: 'Incorrect credentials provided' }, `(${verifyUsernamePassword.name})`);
-                return done(null, false, { message: 'Incorrect credentials provided' });
-            });
-        })
-        .catch(err => {
-            log.error({ err }, `(${verifyUsernamePassword.name})`);
-            return done(err);
-        });
-}
-
-export async function verifyUsernamePasswordAsync(username: string, password: string) {
-    return new Promise<User>((resolve, reject) => {
-        verifyUsernamePassword(username, password, (err: Error, user: User, info: string) => {
-            if (err) {
-                return reject(err);
-            }
-
-            if (user) {
-                return resolve(user);
-            }
-
-            return reject(info);
-        });
-    });
-}
-
-export async function tokenAuthenticate(token: string) {
+export async function tokenAuthenticate(token: string): Promise<User | null> {
     log.info('tokenAuthenticate - jwt: ' + token);
 
     if (process.env.NODE_ENV === 'test') {
@@ -110,9 +32,7 @@ export function sessionAuthenticate() {
             const user = await tokenAuthenticate(req.cookies.jwt);
 
             if (!user) {
-                res.status(401);
-                res.json({ message: 'Invalid token (user not found)' });
-                return next();
+                throw new AuthenticationError('Invalid token (user not found)');
             }
 
             const userReq = req as UserAwareRequest;
